+++ frontend/src/common/components/DesktopNavbar/DesktopNavbar.styles.ts
