+++ frontend/src/test/int/desktopnavbar.int.test.tsx
@@ -1,0 +1,5 @@
+// import { screen, renderWithRouter } from '../utils/test-utils';
+// import '@testing-library/jest-dom/extend-expect';
+
+// TODO: int - That clicking navbar brand takes you to home
+export default {};
